Readme
In the repository we describe the construction and optimization of a CONNIE $1\times5$-specific CNN. We begin human based data-set. Then we apply data augmentation to balance our categories and make the model approximately invariant to random noise. Next, we summarize the model after training and validation process applied with KFold crossvalidation. Finally, we present the classification results, the confusion matrix and the performance of the model.